Feat: Validação do endpoint de criação. -Foi adicionado diversas validações para criação de uma declaração -Foi modificado os DTOS -Incluido codigo no modelmapper para ignorar alguns atributos -Removido endpoint de atualizar declaração, ja que não é nescessário. -Incluido exceções tratadas no response da API.

+++ src/main/java/br/com/futurodev/ReciclaVille_API/model/entity/Material.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -19,5 +21,5 @@ public class Material {
     @Column
     private String nome;
     @Column
-    private Float percentualCompensacao;
+    private BigDecimal percentualCompensacao;
 }
